Refactor component naming for clarity
