+++ src/components/LiveWord.svelte
